pyproject.toml
[tool.poetry]
name = "the-math-guys-bot"
version = "0.1.0"
description = ""
authors = ["MathItYT <benjamin.ubilla@uc.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-dotenv = "^1.0.0"
py-cord = {git = "https://github.com/Pycord-Development/pycord.git", branch="master"}
python-youtube = "^0.9.5"
openai = "^1.43.0"
wolframalpha = "^5.1.3"


[tool.poetry.group.dev.dependencies]
selenium = "^4.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
